Restructure the sample module.
