+++ modules/datacafe-samples/src/main/java/edu/emory/bmi/datacafe/impl/mongo/clinical/main/ExecutorRandomID.java
@@ -13,52 +13,67 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package edu.emory.bmi.datacafe.impl.mysql.clinical.main;
+package edu.emory.bmi.datacafe.impl.mongo.clinical.main;
 
+import edu.emory.bmi.datacafe.constants.MongoConstants;
 import edu.emory.bmi.datacafe.core.CoreExecutorEngine;
 import edu.emory.bmi.datacafe.core.DataSourcesRegistry;
 import edu.emory.bmi.datacafe.hdfs.HdfsConnector;
 import edu.emory.bmi.datacafe.mongo.MongoConnector;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.bson.Document;
 
 import java.util.List;
 
 /**
- * A sample executor for Mongo, where the _id is considered as part of the original data.
+ * The _id is Mongo-generated Random string. Ignore it.
  */
-public class Executor {
-    public static void main(String[] args) {
+public class ExecutorRandomID {
+    private static Logger logger = LogManager.getLogger(ExecutorRandomID.class.getName());
 
+    public static void main(String[] args) {
         CoreExecutorEngine.init();
 
-        String database1 = "clinical";
-        String collection1 = "clinicalData";
+        String database1 = "patients";
+        String collection1 = "patientsData";
 
-        String database2 = "pathology";
-        String collection2 = "pathologyData";
+        String database2 = "admissions";
+        String collection2 = "admissionsData";
 
         DataSourcesRegistry.addDataSource(database1, collection1);
         DataSourcesRegistry.addDataSource(database2, collection2);
 
         // Get the list of IDs from the first data source
-        Document document1 = new Document("Age_at_Initial_Diagnosis", new Document("$gt", 60)).append("Laterality", "Left");
+        Document document1 = new Document("GENDER", "M");
         List ids1 = MongoConnector.getID(database1, collection1, document1);
 
+        if(logger.isDebugEnabled()) {
+            logger.debug("First list of IDs retrieved with ids, " + ids1.size());
+        }
+
         // Get the list of IDs from the second data source
-        Document document2 = new Document("Tumor_Nuclei_Percentage", new Document("$gt", 65));
+        Document document2 = new Document("ADMISSION_TYPE", "EMERGENCY");
         List ids2 = MongoConnector.getID(database2, collection2, document2);
 
-        // Interested Attributes: "patientID", "gender", "laterality"
-        List chosenAttributes1 = MongoConnector.getAttributeValues(database1, collection1, ids1, new String[]{"Gender", "Laterality"});
+        if (logger.isDebugEnabled()) {
+            logger.debug("Second list of IDs retrieved with ids, " + ids2.size());
+        }
 
-        // Interested Attributes: "sliceID", "patientID", "slideBarCode"
-        List chosenAttributes2 = MongoConnector.getAttributeValues(database2, collection2, ids2, new String[]{"BCR_Patient_UID_From_Pathology", "Slide_Barcode"});
+        // ID: "SUBJECT_ID". Other Interested Attributes: "ROW_ID", "GENDER"
+        List chosenAttributes1 = MongoConnector.getAttributeValues(database1, collection1, ids1,
+                new String[]{"SUBJECT_ID", "ROW_ID", "GENDER"}, new String[] {MongoConstants.ID_ATTRIBUTE});
+
+        // ID: "HADM_ID". Other Interested Attributes: "SUBJECT_ID", "DISCHARGE_LOCATION", "MARITAL_STATUS"
+        List chosenAttributes2 = MongoConnector.getAttributeValues(database2, collection2, ids2,
+                new String[]{"HADM_ID", "SUBJECT_ID", "DISCHARGE_LOCATION", "MARITAL_STATUS"},
+                new String[] {MongoConstants.ID_ATTRIBUTE});
 
         List<String>[] attributes = new List[]{chosenAttributes1, chosenAttributes2};
         String[] dataSourcesNames = DataSourcesRegistry.getFullNamesAsArray();
 
         // Write to the Data Lake
         HdfsConnector.writeToWarehouse(dataSourcesNames, attributes);
+
     }
 }
-
