+++ modules/datacafe-samples/src/main/java/edu/emory/bmi/datacafe/impl/mongo/package-info.java
@@ -1,4 +1,4 @@
 /*
  * Clinical1 sample client implementation for Data Cafe, using Mongo DB as the input data source.
  */
-package edu.emory.bmi.datacafe.impl.mysql;
+package edu.emory.bmi.datacafe.impl.mongo;
