+++ modules/datacafe-samples/src/main/java/edu/emory/bmi/datacafe/impl/mysql/clinical/main/package-info.java
@@ -1,0 +1,4 @@
+/*
+ * Clinical1 sample client implementation for Data Cafe, using MySQL as the input data source.
+ */
+package edu.emory.bmi.datacafe.impl.mysql.clinical.main;
